requirements.txt
bandit~=1.7.0
black~=21.0b
flake8~=4.0.0
flask~=2.0.0
logging-formatter-anticrlf==1.2
requests~=2.26.0
asana~=4.0.10
```
